Refactor RPC completion logging tests and remove obsolete policy tests
- Enhanced the RPC completion logging tests by adding comprehensive checks for access logs, including capturing and parsing log entries.
- Introduced new test cases for logging behavior under various scenarios, including error handling and rate limiting.
- Removed the obsolete `rpc_completion_policy_test.go` file as its functionality has been integrated into the main logging tests.
- Improved synchronization and concurrency handling in the tests to ensure accurate logging behavior under load.
