+++ doc/grpc_telemetry.md
@@ -71,29 +71,33 @@ System telemetry in SONiC supports both dial-in mode and dial-out mode. The DB c
 
 # Get and Set audit records
 
-One outer gRPC completion interceptor produces `RPC_ACCESS` records and a
-completion audit decision for every gNMI Get and Set call. Individual
-`GNMI_AUDIT` records use the same glog delivery path and contain `v`,
-`request_id`, `principal`, `auth_result`, `peer_type`, `peer_address`, `method`,
-`path`, `time`, and `code`.
+One outer gRPC completion interceptor emits one `RPC_ACCESS` record for each
+allowed RPC. The unified record contains `v`, `type`, `method`, `peer_type`,
+`peer`, `principal`, `auth_type`, `path`, `code`, `duration_ms`, and
+`suppressed`. `principal`, `auth_type`, and `path` use empty defaults.
 
-Paths contain element names only. Path keys and request and response values are
-not logged. The interceptor uses a verified client-certificate common name or
-SAN as the principal when available, `local` for a Unix socket, and `unknown`
-otherwise. Basic and JWT usernames are not available to the outer interceptor.
-`auth_result` is `denied` for `Unauthenticated` and `PermissionDenied`, `local`
-for a Unix socket, and `not_evaluated` otherwise.
+The logger uses only data already available at interceptor entry or completion
+and never serializes the complete request or response. `principal` copies the presented
+TLS peer certificate Common Name without chain, SAN, role, authentication, or
+authorization validation. After the deferred completion handler passes the
+rate limit, one `deriveRequestFields` wrapper performs the only peer-context
+existence check, then delegates address/type, presented CN, transport auth
+method, and raw path extraction. It is named for the request because paths are
+not peer data. Path keys may be present; request update values and responses
+are not included. Authenticated principal, the existing `TELEMETRY-*` request
+ID, and auth result are deferred until those values are available in the
+interceptor context.
 
-`GNMI_AUDIT` and `RPC_ACCESS` share gRPC-code mapping, peer extraction, JSON
-framing, and the glog line sink. `RPC_ACCESS` retains its independent
-method/code rate limiter and suppression summaries.
+Other unary and streaming methods use the same record. Peer-derived
+`principal` and `auth_type` are populated when available; `path` remains empty
+outside Get/Set. PR #723 method/code rate limiting and suppression summaries
+remain.
 
-Set records are never rate-limited. Get records use three fixed token buckets
-for `ok`, `denied`, and `error` outcomes. Each bucket permits 60 records per
-hour with a burst of 60. A suppressed Get does not reach the glog sink. One
-`GNMI_AUDIT_SUMMARY` record per affected bucket reports the suppressed count
-after the one-hour summary interval without including principal, peer, or path
-data.
+Set records are never rate-limited. Get records use the same method/code bucket
+map and `RPC_ACCESS_SUMMARY` path as other RPCs, with 60 records/hour and burst
+60 instead of PR #723's default one record per 10 seconds. A suppressed Get
+does not reach the glog sink; its common summary reports method, code, and
+suppressed count.
 
 # gRPC operations for system telemetry in SONiC
 As mentioned at the beginning, SONiC gRPC data telemetry is largely based on gNMI protocol,  the GetRquest/GetResponse and SubscribeRequest/SubscribeResponse RPC have been implemented. Since SONiC doesn't have complete YANG data model yet, the DB, TABLE, KEY and Field path hierarchy is used as path to uniquely identify the configuration/state and counter data.
